bugfix: silent death if home exists on multiple devices
Now the process detects this condition early (during ct of static globals), chatters to the console and to the log, and then calls BLAMMOid().  Multiple homes on a single device are allowed, but all homes must be on the same device.  Now a violation results in an easy to understand and see failure so that the administrator can remove the homes on all but wo device.